Literature Review
In this section a selection of papers was used in order to exemine what past researches have conducted in a similar senario. Various papers used different tehcniques in order to check weather penetration testing can be conducted using a Raspberry PI and a drone. Various attacks were discussed to  outline the best solution to be conducted in this disseration