Conclution and Recomendations
The conclution and recomendation section includes a summary of all the findings that were found during the tests conducted. It outlines that the hypotesis question was addressed and verified that it was true.  The strenghts of the results found is that most of the tests conducted were sucessfull and also shows that a drone and a Raspberry PI can be used in order to conduct penetration testing on  a system. Limitations were noted and disscussed.