Results and Discussion
In this document the results that were obtained during the tests were discussed and analysed in detail. Tables and a Graph was used to compare each test more easily and point out various important factors that were evident during the testing phase. The findings found from the results where compared to other researchers paper. It was found that some papers had similar results which confirmed that the results obtained were correct. Furthermore,  the prototype was discussed where it was stated that the two technologies can be integrated and used together. In order to answer the research question in the concluding part of this section it was stated that with various tools it was able to conduct such research possible.